+++ pyfc/__init__.py
@@ -2,7 +2,6 @@
 Feldman-Cousins Frequentist Analysis Framework
 
 Created: v0.1.0 (July 24, 2026)
-Last modified: v0.20.0
 Author: Mauricio Bustamante (mbustamante@gmail.com)
 """